fix query + Integer al posto di Long per il token

+++ progetto-fermi-bean/src/main/java/it/previnet/progettofermi/bean/request/ExampleSearch.java
@@ -4,13 +4,13 @@ import javax.ws.rs.QueryParam;
 
 public class ExampleSearch {
     @QueryParam("token")
-    private Long token;
+    private Integer token;
 
-    public Long getToken() {
+    public Integer getToken() {
         return token;
     }
 
-    public void setToken(Long token) {
+    public void setToken(Integer token) {
         this.token = token;
     }
 }
